Clean-up and fixes for dns
