+++ .github/workflows/run-tests.yml
@@ -33,7 +33,7 @@ jobs:
           sudo microk8s addons repo add community .
           export UNDER_TIME_PRESSURE="True"
           export SKIP_PROMETHEUS="False"
-          sudo -E pytest -s -ra ./tests/test_*.py
+          sudo -E pytest -s -ra ./tests/
           sudo snap remove microk8s --purge
       - name: Running addons tests on strict
         run: |
@@ -48,5 +48,5 @@ jobs:
           export UNDER_TIME_PRESSURE="True"
           export SKIP_PROMETHEUS="False"
           export STRICT="yes"
-          sudo -E pytest -s -ra ./tests/test_*.py
+          sudo -E pytest -s -ra ./tests/
           sudo snap remove microk8s --purge
