+++ tests/test_common.py
@@ -3,36 +3,19 @@ import requests
 import pytest
 import os
 import platform
-import sh
-import yaml
+import subprocess
 
 from utils import (
-    docker,
     get_arch,
-    is_container,
     kubectl,
-    kubectl_get,
     microk8s_disable,
     microk8s_enable,
-    microk8s_reset,
-    run_until_success,
     update_yaml_with_arch,
-    wait_for_installation,
-    wait_for_namespace_termination,
     wait_for_pod_state,
 )
-from subprocess import PIPE, STDOUT, CalledProcessError, check_call, run, check_output
 
 
 class TestCommon(object):
-    @pytest.fixture(scope="session", autouse=True)
-    def clean_up(self):
-        """
-        Clean up after a test
-        """
-        yield
-        microk8s_reset()
-
     @pytest.mark.skipif(
         platform.machine() != "s390x",
         reason="This test is for the limited set of addons s390x has",
@@ -43,6 +26,7 @@ class TestCommon(object):
 
         """
         ip_ranges = "8.8.8.8,1.1.1.1"
+        microk8s_disable("dns")
         print("Enabling DNS")
         microk8s_enable("{}:{}".format("dns", ip_ranges), timeout_insec=500)
         wait_for_pod_state("", "kube-system", "running", label="k8s-app=kube-dns")
@@ -57,6 +41,7 @@ class TestCommon(object):
         """
         ip_ranges = "8.8.8.8,1.1.1.1"
         print("Enabling DNS")
+        microk8s_disable("dns")
         microk8s_enable("{}:{}".format("dns", ip_ranges), timeout_insec=500)
         wait_for_pod_state("", "kube-system", "running", label="k8s-app=kube-dns")
         print("Validating DNS config")
@@ -90,31 +75,6 @@ class TestCommon(object):
         print("Disabling DNS")
         microk8s_disable("dns")
         """
-
-    @pytest.mark.skipif(
-        os.environ.get("STRICT") == "yes",
-        reason="Skipping nfs tests in strict confinement as they are expected to fail",
-    )
-    @pytest.mark.skipif(
-        platform.machine() != "x86_64",
-        reason="NFS tests are only relevant in x86 architectures",
-    )
-    @pytest.mark.skipif(
-        os.environ.get("UNDER_TIME_PRESSURE") == "True",
-        reason="Skipping NFS tests as we are under time pressure",
-    )
-    # NFS addon requires elevated privileges, which fails in lxc due to seccomp.
-    @pytest.mark.skipif(is_container(), reason="NFS tests are skipped in containers")
-    def test_storage_nfs(self):
-        """
-        Sets up and validates NFS Server Provisioner.
-        """
-        print("Enabling NFS")
-        microk8s_enable("nfs")
-        print("Validating NFS")
-        self.validate_storage_nfs()
-        print("Disabling NFS")
-        microk8s_disable("nfs")
 
     def validate_coredns_config(self, nameservers="8.8.8.8,1.1.1.1"):
         """
@@ -158,9 +118,9 @@ class TestCommon(object):
 
     def validate_dashboard_ingress(self):
         """
-        Validate the ingress for dashboard addon by trying to access the kubernetes dashboard
-        using ingress ports. The dashboard will return HTTP 200 and HTML indicating that it is
-        up and running.
+        Validate the ingress for dashboard addon by trying to access the kubernetes 
+        dashboard using ingress ports. The dashboard will return HTTP 200 and HTML 
+        indicating that it is up and running.
         """
         service_ok = False
         attempt = 50
@@ -188,7 +148,7 @@ class TestCommon(object):
         """
         output = kubectl("describe deployment hostpath-provisioner -n kube-system")
         if "hostpath-provisioner-{}:1.0.0".format(get_arch()) in output:
-            # we are running with a hostpath-provisioner that is old and we need to patch it
+            # we are running with an old hostpath-provisioner and we need to patch it
             kubectl(
                 "set image  deployment hostpath-provisioner -n kube-system"
                 " hostpath-provisioner=cdkbot/hostpath-provisioner:1.1.0"
@@ -218,41 +178,6 @@ class TestCommon(object):
                     found = True
         assert found
         assert "myclaim" in output
-        assert "Bound" in output
-        kubectl("delete -f {}".format(manifest))
-
-    def validate_storage_nfs(self):
-        """
-        Validate NFS Storage by creating two Pods mounting the same PVC. (optimal test would be on multinode-cluster)
-        """
-        wait_for_pod_state(
-            "", "nfs-server-provisioner", "running", label="app=nfs-server-provisioner"
-        )
-
-        here = os.path.dirname(os.path.abspath(__file__))
-        manifest = os.path.join(here, "templates", "pvc-nfs.yaml")
-        kubectl("apply -f {}".format(manifest))
-        wait_for_pod_state("", "default", "running", label="app=busybox-pvc-nfs")
-
-        attempt = 50
-        while attempt >= 0:
-            output = kubectl("get pvc -l vol=pvc-nfs")
-            if "Bound" in output:
-                break
-            time.sleep(2)
-            attempt -= 1
-
-        # Make sure the test pod writes data to the storage
-        found = False
-        for root, dirs, files in os.walk("/var/snap/microk8s/common/nfs-storage"):
-            for file in files:
-                if file == "dates1":
-                    found1 = True
-                if file == "dates2":
-                    found2 = True
-        assert found1
-        assert found2
-        assert "pvc-nfs" in output
         assert "Bound" in output
         kubectl("delete -f {}".format(manifest))
