+++ tests/test_keda.py
@@ -1,36 +1,17 @@
 import pytest
 import os
 import platform
-import sh
-import yaml
 
 from utils import (
-    docker,
-    get_arch,
-    is_container,
     kubectl,
-    kubectl_get,
     microk8s_disable,
     microk8s_enable,
-    microk8s_reset,
-    run_until_success,
-    update_yaml_with_arch,
     wait_for_installation,
-    wait_for_namespace_termination,
     wait_for_pod_state,
 )
-from subprocess import PIPE, STDOUT, CalledProcessError, check_call, run, check_output
 
 
 class TestKeda(object):
-    @pytest.fixture(scope="session", autouse=True)
-    def clean_up(self):
-        """
-        Clean up after a test
-        """
-        yield
-        microk8s_reset()
-
     @pytest.mark.skipif(
         platform.machine() != "x86_64",
         reason="KEDA tests are only relevant in x86 architectures",
