+++ tests/test_knative.py
@@ -1,36 +1,18 @@
 import pytest
 import os
 import platform
-import sh
-import yaml
 
 from utils import (
-    docker,
-    get_arch,
-    is_container,
     kubectl,
-    kubectl_get,
     microk8s_disable,
     microk8s_enable,
-    microk8s_reset,
-    run_until_success,
-    update_yaml_with_arch,
     wait_for_installation,
     wait_for_namespace_termination,
     wait_for_pod_state,
 )
-from subprocess import PIPE, STDOUT, CalledProcessError, check_call, run, check_output
 
 
 class TestKnative(object):
-    @pytest.fixture(scope="session", autouse=True)
-    def clean_up(self):
-        """
-        Clean up after a test
-        """
-        yield
-        microk8s_reset()
-
     @pytest.mark.skipif(platform.machine() == "s390x", reason="Not available on s390x")
     @pytest.mark.skipif(
         os.environ.get("UNDER_TIME_PRESSURE") == "True",
@@ -52,7 +34,7 @@ class TestKnative(object):
 
     def validate_knative(self):
         """
-        Validate Knative by deploying the helloworld-go app supports both amd64 and arm64
+        Validate Knative by deploying the helloworld-go app supports both amd64 & arm64
         """
 
         wait_for_installation()
