+++ tests/test_kwasm.py
@@ -1,39 +1,19 @@
 import pytest
 import os
 import platform
-import sh
-import yaml
 
 from utils import (
-    docker,
-    get_arch,
-    is_container,
     kubectl,
-    kubectl_get,
     microk8s_disable,
     microk8s_enable,
-    microk8s_reset,
-    run_until_success,
-    update_yaml_with_arch,
-    wait_for_installation,
-    wait_for_namespace_termination,
     wait_for_pod_state,
 )
-from subprocess import PIPE, STDOUT, CalledProcessError, check_call, run, check_output
 
 
 class TestKwasm(object):
-    @pytest.fixture(scope="session", autouse=True)
-    def clean_up(self):
-        """
-        Clean up after a test
-        """
-        yield
-        microk8s_reset()
-
     @pytest.mark.skipif(
         os.environ.get("STRICT") == "yes",
-        reason="Skipping kwasm tests in strict confinement as they are expected to fail",
+        reason="Skipping kwasm tests in strict confinement as they are expected to fail"
     )
     @pytest.mark.skipif(platform.machine() == "s390x", reason="Not available on s390x")
     def test_kwasm(self):
